Update reward functions
Reward functions no longer clamp on the same number, but have individual clamps to ensure the punishments are logical in size.

Added a reward function that gives an increasing score for moving away from the starting point of the route, which flattens out after a certain distance is reached.

Added evaluation callback to only save best model for each run